+++ src/action_state.rs
@@ -1,6 +1,7 @@
 //! This module contains [`ActionState`] and its supporting methods and impls.
 
 use crate::action_diff::ActionDiff;
+#[cfg(feature = "timing")]
 use crate::timing::Timing;
 use crate::Actionlike;
 use crate::{axislike::DualAxisData, buttonlike::ButtonState};
@@ -8,7 +9,9 @@ use crate::{axislike::DualAxisData, buttonlike::ButtonState};
 use bevy::ecs::component::Component;
 use bevy::prelude::Resource;
 use bevy::reflect::Reflect;
-use bevy::utils::{Duration, HashMap, Instant};
+#[cfg(feature = "timing")]
+use bevy::utils::Duration;
+use bevy::utils::{HashMap, Instant};
 use serde::{Deserialize, Serialize};
 
 /// Metadata about an [`Actionlike`] action
@@ -28,12 +31,47 @@ pub struct ActionData {
     /// The [`DualAxisData`] of the binding that triggered the action.
     pub axis_pair: Option<DualAxisData>,
     /// When was the button pressed / released, and how long has it been held for?
+    #[cfg(feature = "timing")]
     pub timing: Timing,
     /// Was this action consumed by [`ActionState::consume`]?
     ///
     /// Actions that are consumed cannot be pressed again until they are explicitly released.
     /// This ensures that consumed actions are not immediately re-pressed by continued inputs.
     pub consumed: bool,
+    /// Is the action disabled?
+    ///
+    /// While disabled, an action will always report as released, regardless of its actual state.
+    pub disabled: bool,
+}
+
+impl ActionData {
+    /// Is the action currently pressed?
+    #[inline]
+    #[must_use]
+    pub fn pressed(&self) -> bool {
+        !self.disabled && self.state.pressed()
+    }
+
+    /// Was the action pressed since the last time it was ticked?
+    #[inline]
+    #[must_use]
+    pub fn just_pressed(&self) -> bool {
+        !self.disabled && self.state.just_pressed()
+    }
+
+    /// Is the action currently released?
+    #[inline]
+    #[must_use]
+    pub fn released(&self) -> bool {
+        self.disabled || self.state.released()
+    }
+
+    /// Was the action released since the last time it was ticked?
+    #[inline]
+    #[must_use]
+    pub fn just_released(&self) -> bool {
+        !self.disabled && self.state.just_released()
+    }
 }
 
 /// Stores the canonical input-method-agnostic representation of the inputs received
@@ -171,15 +209,16 @@ impl<A: Actionlike> ActionState<A> {
     /// assert!(action_state.pressed(&Action::Jump));
     /// assert!(!action_state.just_pressed(&Action::Jump));
     /// ```
-    pub fn tick(&mut self, current_instant: Instant, previous_instant: Instant) {
+    pub fn tick(&mut self, _current_instant: Instant, _previous_instant: Instant) {
         // Advanced the ButtonState
         self.action_data.values_mut().for_each(|ad| ad.state.tick());
 
-        // Advance the Timings
+        // Advance the Timings if the feature is enabled
+        #[cfg(feature = "timing")]
         self.action_data.values_mut().for_each(|ad| {
             // Durations should not advance while actions are consumed
             if !ad.consumed {
-                ad.timing.tick(current_instant, previous_instant);
+                ad.timing.tick(_current_instant, _previous_instant);
             }
         });
     }
@@ -365,6 +404,7 @@ impl<A: Actionlike> ActionState<A> {
             return;
         }
 
+        #[cfg(feature = "timing")]
         if action_data.state.released() {
             action_data.timing.flip();
         }
@@ -383,11 +423,19 @@ impl<A: Actionlike> ActionState<A> {
         // Once released, consumed actions can be pressed again
         action_data.consumed = false;
 
+        #[cfg(feature = "timing")]
         if action_data.state.pressed() {
             action_data.timing.flip();
         }
 
         action_data.state.release();
+    }
+
+    /// Releases all actions
+    pub fn release_all(&mut self) {
+        for action in self.keys() {
+            self.release(&action);
+        }
     }
 
     /// Consumes the `action`
@@ -436,6 +484,7 @@ impl<A: Actionlike> ActionState<A> {
         // This is the only difference from action_state.release(&action)
         action_data.consumed = true;
         action_data.state.release();
+        #[cfg(feature = "timing")]
         action_data.timing.flip();
     }
 
@@ -447,13 +496,6 @@ impl<A: Actionlike> ActionState<A> {
         }
     }
 
-    /// Releases all actions
-    pub fn release_all(&mut self) {
-        for action in self.keys() {
-            self.release(&action);
-        }
-    }
-
     /// Is this `action` currently consumed?
     #[inline]
     #[must_use]
@@ -461,18 +503,72 @@ impl<A: Actionlike> ActionState<A> {
         matches!(self.action_data(action), Some(action_data) if action_data.consumed)
     }
 
+    /// Disables the `action`
+    #[inline]
+    pub fn disable(&mut self, action: &A) {
+        let action_data = match self.action_data_mut(action) {
+            Some(action_data) => action_data,
+            None => {
+                self.set_action_data(action.clone(), ActionData::default());
+                self.action_data_mut(action).unwrap()
+            }
+        };
+
+        action_data.disabled = true;
+    }
+
+    /// Disables all actions
+    #[inline]
+    pub fn disable_all(&mut self) {
+        for action in self.keys() {
+            self.disable(&action);
+        }
+    }
+
+    /// Is this `action` currently disabled?
+    #[inline]
+    #[must_use]
+    pub fn disabled(&mut self, action: &A) -> bool {
+        match self.action_data(action) {
+            Some(action_data) => action_data.disabled,
+            None => false,
+        }
+    }
+
+    /// Enables the `action`
+    #[inline]
+    pub fn enable(&mut self, action: &A) {
+        let action_data = match self.action_data_mut(action) {
+            Some(action_data) => action_data,
+            None => {
+                self.set_action_data(action.clone(), ActionData::default());
+                self.action_data_mut(action).unwrap()
+            }
+        };
+
+        action_data.disabled = false;
+    }
+
+    /// Enables all actions
+    #[inline]
+    pub fn enable_all(&mut self) {
+        for action in self.keys() {
+            self.enable(&action);
+        }
+    }
+
     /// Is this `action` currently pressed?
     #[inline]
     #[must_use]
     pub fn pressed(&self, action: &A) -> bool {
-        matches!(self.action_data(action), Some(action_data) if action_data.state.pressed())
+        matches!(self.action_data(action), Some(action_data) if action_data.pressed())
     }
 
     /// Was this `action` pressed since the last time [tick](ActionState::tick) was called?
     #[inline]
     #[must_use]
     pub fn just_pressed(&self, action: &A) -> bool {
-        matches!(self.action_data(action), Some(action_data) if action_data.state.just_pressed())
+        matches!(self.action_data(action), Some(action_data) if action_data.just_pressed())
     }
 
     /// Is this `action` currently released?
@@ -482,7 +578,7 @@ impl<A: Actionlike> ActionState<A> {
     #[must_use]
     pub fn released(&self, action: &A) -> bool {
         match self.action_data(action) {
-            Some(action_data) => action_data.state.released(),
+            Some(action_data) => action_data.released(),
             None => true,
         }
     }
@@ -491,7 +587,7 @@ impl<A: Actionlike> ActionState<A> {
     #[inline]
     #[must_use]
     pub fn just_released(&self, action: &A) -> bool {
-        matches!(self.action_data(action), Some(action_data) if action_data.state.just_released())
+        matches!(self.action_data(action), Some(action_data) if action_data.just_released())
     }
 
     #[must_use]
@@ -499,7 +595,7 @@ impl<A: Actionlike> ActionState<A> {
     pub fn get_pressed(&self) -> Vec<A> {
         self.action_data
             .iter()
-            .filter(|(_action, data)| data.state.pressed())
+            .filter(|(_action, data)| data.pressed())
             .map(|(action, _data)| action.clone())
             .collect()
     }
@@ -509,7 +605,7 @@ impl<A: Actionlike> ActionState<A> {
     pub fn get_just_pressed(&self) -> Vec<A> {
         self.action_data
             .iter()
-            .filter(|(_action, data)| data.state.just_pressed())
+            .filter(|(_action, data)| data.just_pressed())
             .map(|(action, _data)| action.clone())
             .collect()
     }
@@ -519,7 +615,7 @@ impl<A: Actionlike> ActionState<A> {
     pub fn get_released(&self) -> Vec<A> {
         self.action_data
             .iter()
-            .filter(|(_action, data)| data.state.released())
+            .filter(|(_action, data)| data.released())
             .map(|(action, _data)| action.clone())
             .collect()
     }
@@ -529,7 +625,7 @@ impl<A: Actionlike> ActionState<A> {
     pub fn get_just_released(&self) -> Vec<A> {
         self.action_data
             .iter()
-            .filter(|(_action, data)| data.state.just_released())
+            .filter(|(_action, data)| data.just_released())
             .map(|(action, _data)| action.clone())
             .collect()
     }
@@ -544,6 +640,7 @@ impl<A: Actionlike> ActionState<A> {
     /// that corresponds exactly to the start of a frame, rather than relying on idiosyncratic timing.
     ///
     /// This will also be [`None`] if the action was never pressed or released.
+    #[cfg(feature = "timing")]
     pub fn instant_started(&self, action: &A) -> Option<Instant> {
         let action_data = self.action_data(action)?;
         action_data.timing.instant_started
@@ -552,6 +649,7 @@ impl<A: Actionlike> ActionState<A> {
     /// The [`Duration`] for which the action has been held or released
     ///
     /// This will be [`Duration::ZERO`] if the action was never pressed or released.
+    #[cfg(feature = "timing")]
     pub fn current_duration(&self, action: &A) -> Duration {
         self.action_data(action)
             .map(|data| data.timing.current_duration)
@@ -564,6 +662,7 @@ impl<A: Actionlike> ActionState<A> {
     /// the action was last pressed or released.
     ///
     /// This will be [`Duration::ZERO`] if the action was never pressed or released.
+    #[cfg(feature = "timing")]
     pub fn previous_duration(&self, action: &A) -> Duration {
         self.action_data(action)
             .map(|data| data.timing.previous_duration)