+++ RELEASES.md
@@ -14,6 +14,10 @@
 - removed `InputMap::insert_chord` and `InputMap::insert_modified` due to their limited applicability within the type system.
   - the new `InputChord` contructors and builders allow you to define chords with guaranteed type safety.
   - the new `ModifierKey::with` method simplifies the creation of input chords that include the modifier and your desired input.
+- the `timing` field of the `ActionData` is now disabled by default. Timing information will only be collected
+  if the `timing` feature is enabled. It is disabled by default because most games don't require timing information.
+  (how long a button was pressed for)
+- removed `ToggleActions` resource in favor of new methods on `ActionState`: `disable_all`, `disable(action)`, `enable_all`, `enable(action)`, and `disabled(action)`.
 - removed `InputMap::build` method in favor of new fluent builder pattern (see 'Usability: InputMap' for details).
 - renamed `InputMap::which_pressed` method to `process_actions` to better reflect its current functionality for clarity.
 - removed `DeadZoneShape` in favor of new dead zone processors (see 'Enhancements: Input Processors' for details).
@@ -152,6 +156,7 @@ Input processors allow you to create custom logic for axis-like input manipulati
 
 ### Bugs
 
+- fixed a bug where enabling a pressed action would read as `just_pressed`, and disabling a pressed action would read as `just_released`.
 - fixed a bug in `InputStreams::button_pressed()` where unrelated gamepads were not filtered out when an `associated_gamepad` is defined.
 
 ## Version 0.13.3